pyproject.toml
[tool.poetry]
name = "todoink"
version = "0.1.0"
description = ""
authors = ["Lorenzo Farinelli <lorenzo.farinelli@outlook.it>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
todoist-api-python = "^2.1.3"
flask = "^3.0.0"
flask-cors = "^4.0.0"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
